+++ android/src/main/kotlin/com/app/helpshift_wrapper/Constants.kt
@@ -1,0 +1,12 @@
+package com.app.helpshift_wrapper
+
+object Constants {
+    const val methodConfigureSdk="configure_sdk"
+    const val allConversation="all_conversation"
+    const val allFAQs="all_faq" // which are published
+    const val methodSingleFAQsSection="single_faq_s"
+    const val methodSingleFAQDetail="single_faq_detail"
+    const val methodLogin="login"
+    const val methodLogout="logout"
+    const val methodSetLanguage="set_language"
+}