+++ lib/native_method_constants.dart
@@ -1,0 +1,10 @@
+class Methods {
+  static var methodConfigureSdk = "configure_sdk";
+  static var allConversation = "all_conversation";
+  static var allFAQs = "all_faq"; // which are published
+  static var methodSingleFAQsSection = "single_faq_s";
+  static var methodSingleFAQDetail = "single_faq_detail";
+  static var methodLogin = "login";
+  static var methodLogout = "logout";
+  static var methodSetLanguage = "set_language";
+}