+++ CHANGELOG.md
@@ -1,3 +1,25 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Have all the methods to see FAQ's section with custom configMap and sectionId and publishId
+
+## 0.0.2
+* Updated Few Files
+
+## 0.0.3
+* removed unnecessary plugin
+
+## 0.0.4
+* Helpshift SDK version upgraded
+
+## 0.0.5
+* Custom Issue Fields Added,
+* Home and Github Repository URL provided
+
+## 0.0.6
+* Login and Logout Functionalities added
+
+## 0.0.7
+* Login and Logout Functionalities updated
+
+## 0.0.8
+* Custom Metadata updated